pyproject.toml
[tool.poetry]
name = "qukit"
version = "0.1.0"
description = ""
authors = ["GBisi <giuseppe.bisicchia@phd.unipi.it>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
frozendict = "^2.4.4"
pytket = "^1.28.0"
qiskit = "<1.1.0"
qiskit-ionq = "^0.5.2"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
qiskit-ibm-runtime = "^0.23.0"
qiskit-aer = "^0.14.2"
click = "^8.1.7"
pytest-timeout = "^2.3.1"
pennylane = "^0.36.0"
pennylane-qiskit = "^0.36.0"
kahypar = "^1.3.5"
cirq = "^1.4.0"
ply = "^3.11"
pyquil = "^4.11.0"
cirq-rigetti = "^1.4.0"
qbraid = "^0.6.1"
boto3 = "^1.34.131"
amazon-braket-sdk = "^1.81.1"
pylatexenc = "^2.10"
path = "^16.14.0"


[tool.poetry.group.dev.dependencies]
commitizen = "^3.27.0"
pre-commit = "^3.7.1"
pytest = "^8.2.1"
pytest-cov = "^5.0.0"
pylint = "^3.2.2"
pytest-xdist = "^3.6.1"
pytest-testmon = "^2.1.1"
ipykernel = "^6.29.4"
mqt-bench = "^1.1.3"
tqdm = "^4.66.4"
notebook = "^7.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 110

[tool.flake8]
per-file-ignores = [
    '__init__.py:F401',
]
max-line-length = 110

[tool.isort]
profile = "black"
line_length = 110

[tool.mypy]
ignore_missing_imports = false
strict = true
no_implicit_reexport = false

[tool.numpydoc_validation]
checks = [
    "all",  # report on all checks
    "GL01", # "Docstring text (summary) should start in the line immediately after the opening quotes (not in the same line, or leaving a blank line in between)
    "ES01", # but don't require an extended summary
    "EX01", # or examples
    "SA01", # or a see also section
    "SS06", # and don't require the summary to fit on one line
]
override_SS05 = [  # allow docstrings to start with these words
    '^Process ',
    '^Assess ',
    '^Access ',
]

[tool.docformatter]
recursive = true
wrap-summaries = 110
wrap-descriptions = 110

[tool.pylint]
max-line-length=110
